Opprydding i saksflyt-tester (integrasjonstester)
