+++ src/test/kotlin/no/nav/eessi/pensjon/integrasjonstest/saksflyt/PBuc05IntegrationTest.kt
@@ -1,39 +1,51 @@
 package no.nav.eessi.pensjon.integrasjonstest.saksflyt
 
+import io.mockk.every
+import io.mockk.mockk
+import io.mockk.slot
+import io.mockk.verify
+import no.nav.eessi.pensjon.handler.OppgaveMelding
+import no.nav.eessi.pensjon.json.mapJsonToAny
+import no.nav.eessi.pensjon.json.typeRefs
+import no.nav.eessi.pensjon.models.*
 import no.nav.eessi.pensjon.models.Enhet.AUTOMATISK_JOURNALFORING
 import no.nav.eessi.pensjon.models.Enhet.ID_OG_FORDELING
 import no.nav.eessi.pensjon.models.Enhet.NFP_UTLAND_AALESUND
 import no.nav.eessi.pensjon.models.Enhet.PENSJON_UTLAND
 import no.nav.eessi.pensjon.models.Enhet.UFORE_UTLAND
 import no.nav.eessi.pensjon.models.Enhet.UFORE_UTLANDSTILSNITT
-import no.nav.eessi.pensjon.models.SakInformasjon
-import no.nav.eessi.pensjon.models.SakStatus
 import no.nav.eessi.pensjon.models.Tema.PENSJON
 import no.nav.eessi.pensjon.models.Tema.UFORETRYGD
-import no.nav.eessi.pensjon.models.YtelseType
+import no.nav.eessi.pensjon.models.sed.DocStatus
+import no.nav.eessi.pensjon.models.sed.Document
 import no.nav.eessi.pensjon.models.sed.Rolle
+import no.nav.eessi.pensjon.models.sed.SED
+import no.nav.eessi.pensjon.personoppslag.pdl.model.NorskIdent
+import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 
-@DisplayName("P_BUC_05 - Utgående journalføring - IntegrationTest")
+@DisplayName("P_BUC_05 - IntegrationTest")
 internal class PBuc05IntegrationTest : JournalforingTestBase() {
 
     /**
-     * P_BUC_05 DEL 1
+     * Flytskjema utgående:
+     * https://confluence.adeo.no/pages/viewpage.action?pageId=387092731
      *
-     * Flytskjema (utgående):
-     *  https://confluence.adeo.no/pages/viewpage.action?pageId=387092731
+     * Flytskjema inngående:
+     * https://confluence.adeo.no/pages/viewpage.action?pageId=387092731
      */
 
 
     /* ============================
-     * SCENARIO 1
+     * UTGÅENDE
      * ============================ */
 
     @Nested
-    inner class Scenario1 {
+    @DisplayName("Utgående - Scenario 1")
+    inner class Scenario1Utgaende {
         @Test
         fun `1 person i SED fnr finnes men ingen bestemsak Så journalføres på ID_OG_FORDELING`() {
             testRunner(FNR_OVER_60, saker = emptyList(), sakId = SAK_ID) {
@@ -61,12 +73,9 @@ internal class PBuc05IntegrationTest : JournalforingTestBase() {
     }
 
 
-    /* ============================
-     * SCENARIO 2
-     * ============================ */
-
     @Nested
-    inner class Scenario2 {
+    @DisplayName("Utgående - Scenario 2")
+    inner class Scenario2Utgaende {
         @Test
         fun `2 personer i SED, fnr og rolle mangler, men saksnummer finnes`() {
             testRunnerFlerePersoner(fnr = null, fnrAnnenPerson = null, rolle = Rolle.ETTERLATTE, sakId = SAK_ID) {
@@ -101,12 +110,9 @@ internal class PBuc05IntegrationTest : JournalforingTestBase() {
     }
 
 
-    /* ============================
-     * SCENARIO 3
-     * ============================ */
-
     @Nested
-    inner class Scenario3 {
+    @DisplayName("Utgående - Scenario 3")
+    inner class Scenario3Utgaende {
         @Test
         fun `1 person i SED fnr finnes, saktype er GENRL, men finnes flere sakstyper`() {
             val saker = listOf(
@@ -197,12 +203,9 @@ internal class PBuc05IntegrationTest : JournalforingTestBase() {
     }
 
 
-    /* ============================
-     * SCENARIO 4
-     * ============================ */
-
     @Nested
-    inner class Scenario4 {
+    @DisplayName("Utgående - Scenario 4")
+    inner class Scenario4Utgaende {
         @Test
         fun `1 person i SED fnr finnes og saktype er GENRL, med flere sakstyper, person bosatt Norge`() {
             val saker = listOf(
@@ -278,12 +281,9 @@ internal class PBuc05IntegrationTest : JournalforingTestBase() {
     }
 
 
-    /* ============================
-     * SCENARIO 5
-     * ============================ */
-
     @Nested
-    inner class Scenario5 {
+    @DisplayName("Utgående - Scenario 5")
+    inner class Scenario5Utgaende {
         @Test
         fun `1 person i SED fnr finnes og bestemsak finner sak ALDER Så journalføres automatisk på tema PENSJON`() {
             val saker = listOf(
@@ -319,12 +319,9 @@ internal class PBuc05IntegrationTest : JournalforingTestBase() {
     }
 
 
-    /* ============================
-     * SCENARIO 6
-     * ============================ */
-
     @Nested
-    inner class Scenario6 {
+    @DisplayName("Utgående - Scenario 6")
+    inner class Scenario6Utgaende {
         @Test
         fun `2 personer i SED, har rolle GJENLEV, fnr finnes, mangler sakInformasjon`() {
             val fnrAnnenPerson = FNR_BARN
@@ -396,12 +393,9 @@ internal class PBuc05IntegrationTest : JournalforingTestBase() {
     }
 
 
-    /* ============================
-     * SCENARIO 7
-     * ============================ */
-
     @Nested
-    inner class Scenario7 {
+    @DisplayName("Utgående - Scenario 7")
+    inner class Scenario7Utgaende {
         @Test
         fun `2 personer i SED, har rolle familiemedlem, fnr finnes og bestemsak finner sak UFØRE Så journalføres automatisk på tema UFORETRYGD`() {
             val saker = listOf(
@@ -456,12 +450,9 @@ internal class PBuc05IntegrationTest : JournalforingTestBase() {
     }
 
 
-    /* ============================
-     * SCENARIO 8
-     * ============================ */
-
     @Nested
-    inner class Scenario8 {
+    @DisplayName("Utgående - Scenario 8")
+    inner class Scenario8Utgaende {
         @Test
         fun `2 personer i SED, har rolle barn 03 og sak er ALDER`() {
             val saker = listOf(
@@ -536,12 +527,9 @@ internal class PBuc05IntegrationTest : JournalforingTestBase() {
     }
 
 
-    /* ============================
-     * SCENARIO 9
-     * ============================ */
-
     @Nested
-    inner class Scenario9 {
+    @DisplayName("Utgående - Scenario 9")
+    inner class Scenario9Utgaende {
         @Test
         fun `2 personer i SED fnr finnes og rolle er barn, og saktype er BARNEP`() {
             val saker = listOf(
@@ -667,4 +655,915 @@ internal class PBuc05IntegrationTest : JournalforingTestBase() {
             }
         }
     }
+
+    @Nested
+    @DisplayName("Utgående - Scenario 13")
+    inner class Scenario13Utgaende {
+        @Test
+        fun `2 personer angitt, gyldig fnr og ufgyldig fnr annenperson, rolle er 01, bosatt Norge del 4`() {
+            val sed = createSed(SedType.P8000,
+                FNR_OVER_60, createAnnenPerson(fnr = FNR_BARN, rolle = Rolle.ETTERLATTE), null)
+            every { euxKlient.hentSed(any(), any()) } returns sed
+
+            every { fagmodulKlient.hentAlleDokumenter(any()) } returns getMockDocuments()
+            every { euxKlient.hentSedDokumenter(any(), any()) } returns getResource("/pdf/pdfResponseUtenVedlegg.json")
+
+            val voksen = createBrukerWith(FNR_OVER_60, "Voksen", "Vanlig", "NOR", "1213", aktorId = AKTOER_ID)
+            every { personService.hentPerson(NorskIdent(FNR_OVER_60)) } returns voksen
+            every { personService.hentPerson(NorskIdent(FNR_BARN)) } returns null
+
+            val hendelse = createHendelseJson(SedType.P8000)
+
+            val meldingSlot = slot<String>()
+            every { oppgaveHandlerKafka.sendDefault(any(), capture(meldingSlot)).get() } returns mockk()
+
+            val (journalpost, _) = initJournalPostRequestSlot()
+
+            listener.consumeSedSendt(hendelse, mockk(relaxed = true), mockk(relaxed = true))
+
+            val oppgaveMelding = mapJsonToAny(meldingSlot.captured, typeRefs<OppgaveMelding>())
+
+            assertEquals("JOURNALFORING", oppgaveMelding.oppgaveType())
+
+            val request = journalpost.captured
+            assertEquals(PENSJON, request.tema)
+            assertEquals(ID_OG_FORDELING, request.journalfoerendeEnhet)
+            Assertions.assertNull(request.bruker)
+
+            verify(exactly = 1) { fagmodulKlient.hentAlleDokumenter(any()) }
+            verify(exactly = 1) { euxKlient.hentSed(any(), any()) }
+        }
+
+        @Test
+        fun `Hente opp korrekt fnr fra P8000 som er sendt fra oss med flere P8000 i BUC`() {
+            val fnr = "28127822044"
+            val afnr = "05127921999"
+            val aktoera = "${fnr}1111"
+            val aktoerf = "${fnr}0000"
+            val saknr = "1223123123"
+
+            val sedP8000_2 = createSed(SedType.P8000, fnr, createAnnenPerson(fnr = afnr, rolle = Rolle.ETTERLATTE), saknr)
+            val sedP8000sendt = createSed(SedType.P8000, fnr, createAnnenPerson(fnr = afnr, rolle = Rolle.ETTERLATTE), saknr)
+            val sedP8000recevied = createSed(SedType.P8000, null, createAnnenPerson(fnr = null, rolle = Rolle.ETTERLATTE), null)
+
+            val dokumenter = mapJsonToAny(getResource("/fagmodul/alldocumentsids_P_BUC_05_multiP8000.json"), typeRefs<List<Document>>())
+            every { fagmodulKlient.hentAlleDokumenter(any()) } returns dokumenter
+            every { euxKlient.hentSed(any(), any()) } returns sedP8000_2 andThen sedP8000recevied andThen sedP8000sendt
+            every { personService.harAdressebeskyttelse(any(), any()) } returns false
+            every { euxKlient.hentSedDokumenter(any(), any()) } returns getResource("/pdf/pdfResponseUtenVedlegg.json")
+            every { personService.hentPerson(NorskIdent(afnr)) } returns createBrukerWith(afnr, "Lever", "Helt i live", "NOR", aktorId = aktoera)
+            every { personService.hentPerson(NorskIdent(fnr)) } returns createBrukerWith(fnr, "Død", "Helt Død", "NOR", aktorId = aktoerf)
+
+            val saker = listOf(
+                SakInformasjon(sakId = "34234234", sakType = YtelseType.ALDER, sakStatus = SakStatus.AVSLUTTET),
+                SakInformasjon(sakId = saknr, sakType = YtelseType.UFOREP, sakStatus = SakStatus.LOPENDE),
+                SakInformasjon(sakId = "34234123", sakType = YtelseType.GENRL, sakStatus = SakStatus.AVSLUTTET)
+            )
+            every { fagmodulKlient.hentPensjonSaklist(aktoera) } returns saker
+            every { journalpostKlient.oppdaterDistribusjonsinfo(any()) } returns Unit
+
+            val (journalpost, _) = initJournalPostRequestSlot(true)
+            val hendelse = createHendelseJson(SedType.P6000, BucType.P_BUC_05)
+
+            val meldingSlot = slot<String>()
+            every { oppgaveHandlerKafka.sendDefault(any(), capture(meldingSlot)).get() } returns mockk()
+
+            listener.consumeSedSendt(hendelse, mockk(relaxed = true), mockk(relaxed = true))
+
+            val request = journalpost.captured
+
+            // forvent tema == PEN og enhet 9999
+            assertEquals(UFORETRYGD, request.tema)
+            assertEquals(AUTOMATISK_JOURNALFORING, request.journalfoerendeEnhet)
+            assertEquals(afnr, request.bruker?.id)
+
+            verify(exactly = 1) { fagmodulKlient.hentAlleDokumenter(any()) }
+            verify(exactly = 1) { fagmodulKlient.hentPensjonSaklist(any()) }
+            verify(exactly = 3) { euxKlient.hentSed(any(), any()) }
+
+        }
+
+        @Test
+        fun `Scenario 13 - 0 Sed sendes som svar med flere personer pa tidligere mottatt P8000, opprettes en journalføringsoppgave på tema PEN og enhet ID OG FORDELING `() {
+            val sedP8000recevied = createSed(SedType.P8000, null, fdato = "1955-07-11")
+            val sedP5000sent = createSedPensjon(SedType.P5000,
+                FNR_OVER_60, gjenlevendeFnr = FNR_BARN
+            )
+
+            val alleDocumenter = listOf(
+                Document("10001", SedType.P8000, DocStatus.RECEIVED),
+                Document("30002", SedType.P5000, DocStatus.SENT)
+            )
+
+            every { personService.hentPerson(NorskIdent(FNR_BARN)) } returns createBrukerWith(FNR_BARN, "Lever", "Helt i live", "NOR", aktorId = AKTOER_ID)
+            every { personService.hentPerson(NorskIdent(FNR_OVER_60)) } returns createBrukerWith(FNR_OVER_60, "Død", "Helt Død", "NOR", aktorId = AKTOER_ID_2)
+
+            every { fagmodulKlient.hentAlleDokumenter(any()) } returns alleDocumenter
+            every { euxKlient.hentSed(any(), any()) } returns sedP8000recevied andThen sedP5000sent
+            every { euxKlient.hentSedDokumenter(any(), any()) } returns getResource("/pdf/pdfResponseUtenVedlegg.json")
+
+            val meldingSlot = slot<String>()
+            every { oppgaveHandlerKafka.sendDefault(any(), capture(meldingSlot)).get() } returns mockk()
+            val (journalpost, journalpostResponse) = initJournalPostRequestSlot()
+            val hendelse = createHendelseJson(SedType.P5000, BucType.P_BUC_05)
+
+            listener.consumeSedSendt(hendelse, mockk(relaxed = true), mockk(relaxed = true))
+
+            val request = journalpost.captured
+            val oppgaveMelding = mapJsonToAny(meldingSlot.captured, typeRefs<OppgaveMelding>())
+
+            assertEquals("JOURNALFORING", oppgaveMelding.oppgaveType())
+            assertEquals(ID_OG_FORDELING, oppgaveMelding.tildeltEnhetsnr)
+            assertEquals(journalpostResponse.journalpostId, oppgaveMelding.journalpostId)
+            assertEquals("P5000", oppgaveMelding.sedType?.name)
+
+            assertEquals("UTGAAENDE", request.journalpostType.name)
+            assertEquals(PENSJON, request.tema)
+            assertEquals(ID_OG_FORDELING, request.journalfoerendeEnhet)
+
+            verify(exactly = 1) { fagmodulKlient.hentAlleDokumenter(any()) }
+            verify(exactly = 2) { euxKlient.hentSed(any(), any()) }
+        }
+
+        @Test
+        fun `Scenario 13 - 1 Sed sendes som svar med flere personer pa tidligere mottatt P8000, opprettes en journalføringsoppgave på tema PEN og enhet ID OG FORDELING `() {
+            val sedP8000recevied = createSed(SedType.P8000, null, fdato = "1955-07-11")
+            val sedP5000sent = createSed(SedType.P5000, null, fdato = "1955-07-11")
+
+            val alleDocumenter = listOf(
+                Document("10001", SedType.P8000, DocStatus.RECEIVED),
+                Document("30002", SedType.P5000, DocStatus.SENT)
+            )
+
+            every { fagmodulKlient.hentAlleDokumenter(any()) } returns alleDocumenter
+            every { euxKlient.hentSed(any(), any()) } returns sedP8000recevied andThen sedP5000sent
+            every { euxKlient.hentSedDokumenter(any(), any()) } returns getResource("/pdf/pdfResponseUtenVedlegg.json")
+
+            val meldingSlot = slot<String>()
+            every { oppgaveHandlerKafka.sendDefault(any(), capture(meldingSlot)).get() } returns mockk()
+            val (journalpost, journalpostResponse) = initJournalPostRequestSlot()
+            val hendelse = createHendelseJson(SedType.P5000, BucType.P_BUC_05)
+
+            listener.consumeSedSendt(hendelse, mockk(relaxed = true), mockk(relaxed = true))
+
+            val request = journalpost.captured
+            val oppgaveMelding = mapJsonToAny(meldingSlot.captured, typeRefs<OppgaveMelding>())
+
+            assertEquals("JOURNALFORING", oppgaveMelding.oppgaveType())
+            assertEquals(ID_OG_FORDELING, oppgaveMelding.tildeltEnhetsnr)
+            assertEquals(journalpostResponse.journalpostId, oppgaveMelding.journalpostId)
+            assertEquals("P5000", oppgaveMelding.sedType?.name)
+
+            assertEquals("UTGAAENDE", request.journalpostType.name)
+            assertEquals(PENSJON, request.tema)
+            assertEquals(ID_OG_FORDELING, request.journalfoerendeEnhet)
+
+            verify(exactly = 1) { fagmodulKlient.hentAlleDokumenter(any()) }
+            verify(exactly = 2) { euxKlient.hentSed(any(), any()) }
+        }
+
+        @Test
+        fun `Scenario 13 - 2 Sed sendes som svar med fnr pa tidligere mottatt P8000, opprettes en journalføringsoppgave på tema PEN og enhet ID OG FORDELING `() {
+            val fnr = "07115521999"
+            val aktoer = "${fnr}111"
+            val sedP8000recevied = createSed(SedType.P8000, null, fdato = "1955-07-11")
+            val sedP5000sent = createSed(SedType.P5000, fnr)
+
+            val alleDocumenter = listOf(
+                Document("10001", SedType.P8000, DocStatus.RECEIVED),
+                Document("30002", SedType.P5000, DocStatus.SENT)
+            )
+
+            every { fagmodulKlient.hentAlleDokumenter(any()) } returns alleDocumenter
+            every { euxKlient.hentSed(any(), any()) } returns sedP8000recevied andThen sedP5000sent
+            every { euxKlient.hentSedDokumenter(any(), any()) } returns getResource("/pdf/pdfResponseUtenVedlegg.json")
+
+            every { personService.hentPerson(NorskIdent(fnr)) } returns createBrukerWith(fnr, "Lever", "Helt i live", "NOR", aktorId = aktoer)
+
+            val meldingSlot = slot<String>()
+            every { oppgaveHandlerKafka.sendDefault(any(), capture(meldingSlot)).get() } returns mockk()
+            val (journalpost, journalpostResponse) = initJournalPostRequestSlot()
+            val hendelse = createHendelseJson(SedType.P5000, BucType.P_BUC_05)
+
+            listener.consumeSedSendt(hendelse, mockk(relaxed = true), mockk(relaxed = true))
+
+            val request = journalpost.captured
+            val oppgaveMelding = mapJsonToAny(meldingSlot.captured, typeRefs<OppgaveMelding>())
+
+            assertEquals("JOURNALFORING", oppgaveMelding.oppgaveType())
+            assertEquals(ID_OG_FORDELING, oppgaveMelding.tildeltEnhetsnr)
+            assertEquals(journalpostResponse.journalpostId, oppgaveMelding.journalpostId)
+            assertEquals("P5000", oppgaveMelding.sedType?.name)
+
+            assertEquals("UTGAAENDE", request.journalpostType.name)
+            assertEquals(PENSJON, request.tema)
+            assertEquals(ID_OG_FORDELING, request.journalfoerendeEnhet)
+
+            verify(exactly = 1) { fagmodulKlient.hentAlleDokumenter(any()) }
+            verify(exactly = 2) { euxKlient.hentSed(any(), any()) }
+
+        }
+
+        @Test
+        fun `Scenario 13 - 2 Sed sendes som svar med fnr pa tidligere mottatt P8000 ingen ident, svar sed med fnr og sakid i sed journalføres automatisk `() {
+            val fnr = FNR_VOKSEN
+            val aktoer = "${fnr}111"
+            val sakid = SAK_ID
+            val sedP8000recevied = createSed(SedType.P8000, null, fdato = "1955-07-11")
+            val sedP9000sent = createSed(SedType.P9000, fnr, eessiSaknr = sakid)
+
+            val alleDocumenter = listOf(
+                Document("10001", SedType.P8000, DocStatus.RECEIVED),
+                Document("30002", SedType.P9000, DocStatus.SENT)
+            )
+
+            val saker = listOf(
+                SakInformasjon(sakId = "34234234", sakType = YtelseType.OMSORG, sakStatus = SakStatus.LOPENDE),
+                SakInformasjon(sakId = "23232312", sakType = YtelseType.GENRL, sakStatus = SakStatus.AVSLUTTET),
+                SakInformasjon(sakId = sakid, sakType = YtelseType.UFOREP, sakStatus = SakStatus.LOPENDE)
+            )
+
+            every { fagmodulKlient.hentPensjonSaklist(aktoer) } returns saker
+            every { fagmodulKlient.hentAlleDokumenter(any()) } returns alleDocumenter
+            every { euxKlient.hentSed(any(), any()) } returns sedP8000recevied andThen sedP9000sent
+            every { euxKlient.hentSedDokumenter(any(), any()) } returns getResource("/pdf/pdfResponseUtenVedlegg.json")
+
+            every { personService.hentPerson(NorskIdent(fnr)) } returns createBrukerWith(fnr, "KRAFTIG ", "VEGGPRYD", "NOR", aktorId = aktoer)
+
+            val meldingSlot = slot<String>()
+            every { oppgaveHandlerKafka.sendDefault(any(), capture(meldingSlot)).get() } returns mockk()
+            val (journalpost, journalpostResponse) = initJournalPostRequestSlot()
+
+            val hendelse = createHendelseJson(SedType.P9000, BucType.P_BUC_05)
+
+            listener.consumeSedSendt(hendelse, mockk(relaxed = true), mockk(relaxed = true))
+
+            val request = journalpost.captured
+            val oppgaveMelding = mapJsonToAny(meldingSlot.captured, typeRefs<OppgaveMelding>())
+
+            assertEquals("JOURNALFORING", oppgaveMelding.oppgaveType())
+            assertEquals(AUTOMATISK_JOURNALFORING, oppgaveMelding.tildeltEnhetsnr)
+            assertEquals(journalpostResponse.journalpostId, oppgaveMelding.journalpostId)
+            assertEquals("P9000", oppgaveMelding.sedType?.name)
+
+            assertEquals("UTGAAENDE", request.journalpostType.name)
+            assertEquals(UFORETRYGD, request.tema)
+            assertEquals(AUTOMATISK_JOURNALFORING, request.journalfoerendeEnhet)
+            assertEquals(fnr, request.bruker?.id!!)
+
+            verify(exactly = 1) { fagmodulKlient.hentAlleDokumenter(any()) }
+            verify(exactly = 2) { euxKlient.hentSed(any(), any()) }
+        }
+
+        @Test
+        fun `Scenario 13 - 2 Sed sendes som svar med fnr pa tidligere mottatt P8000 ingen ident, svar sed med fnr og ingen sakid i sed journalføres UFO `() {
+            val fnr = FNR_VOKSEN
+            val aktoer = "${fnr}111"
+            val sakid = SAK_ID
+            val sedP8000recevied = createSed(SedType.P8000, null, fdato = "1955-07-11")
+            val sedP9000sent = createSed(SedType.P9000, fnr, eessiSaknr = sakid)
+
+            val alleDocumenter = listOf(
+                Document("10001", SedType.P8000, DocStatus.RECEIVED),
+                Document("30002", SedType.P9000, DocStatus.SENT)
+            )
+
+            val saker = listOf(
+                SakInformasjon(sakId = "34234234", sakType = YtelseType.OMSORG, sakStatus = SakStatus.LOPENDE),
+                SakInformasjon(sakId = "23232312", sakType = YtelseType.GENRL, sakStatus = SakStatus.AVSLUTTET),
+                SakInformasjon(sakId = "123123123123123", sakType = YtelseType.UFOREP, sakStatus = SakStatus.LOPENDE)
+            )
+
+            every { fagmodulKlient.hentPensjonSaklist(aktoer) } returns saker
+            every { fagmodulKlient.hentAlleDokumenter(any()) } returns alleDocumenter
+            every { euxKlient.hentSed(any(), any()) } returns sedP8000recevied andThen sedP9000sent
+            every { euxKlient.hentSedDokumenter(any(), any()) } returns getResource("/pdf/pdfResponseUtenVedlegg.json")
+
+            every { personService.hentPerson(NorskIdent(fnr)) } returns createBrukerWith(fnr, "KRAFTIG ", "VEGGPRYD", "NOR", aktorId = aktoer)
+
+            val meldingSlot = slot<String>()
+            every { oppgaveHandlerKafka.sendDefault(any(), capture(meldingSlot)).get() } returns mockk()
+            val (journalpost, journalpostResponse) = initJournalPostRequestSlot()
+
+            val hendelse = createHendelseJson(SedType.P9000, BucType.P_BUC_05)
+
+            listener.consumeSedSendt(hendelse, mockk(relaxed = true), mockk(relaxed = true))
+
+            val request = journalpost.captured
+            val oppgaveMelding = mapJsonToAny(meldingSlot.captured, typeRefs<OppgaveMelding>())
+
+            assertEquals("JOURNALFORING", oppgaveMelding.oppgaveType())
+            assertEquals(ID_OG_FORDELING, oppgaveMelding.tildeltEnhetsnr)
+            assertEquals(journalpostResponse.journalpostId, oppgaveMelding.journalpostId)
+            assertEquals("P9000", oppgaveMelding.sedType?.name)
+
+            assertEquals("UTGAAENDE", request.journalpostType.name)
+            assertEquals(PENSJON, request.tema)
+            assertEquals(ID_OG_FORDELING, request.journalfoerendeEnhet)
+            assertEquals(fnr, request.bruker?.id!!)
+
+            verify(exactly = 1) { fagmodulKlient.hentAlleDokumenter(any()) }
+            verify(exactly = 2) { euxKlient.hentSed(any(), any()) }
+
+        }
+
+        @Test
+        fun `Scenario 13 - 3 Sed sendes som svar med fnr og sak finnes og er GENRL pa tidligere mottatt P8000, opprettes en journalføringsoppgave på tema NFP UTLAND AALESUND`() {
+            val fnr = FNR_VOKSEN
+            val sakid = "1231232323"
+            val aktoer = "${fnr}111"
+            val sedP8000recevied = createSed(SedType.P8000, null, fdato = "1955-07-11")
+            val sedP5000sent = createSed(SedType.P5000, fnr, eessiSaknr = sakid)
+
+            val alleDocumenter = listOf(
+                Document("10001", SedType.P8000, DocStatus.RECEIVED),
+                Document("30002", SedType.P5000, DocStatus.SENT)
+            )
+
+            every { fagmodulKlient.hentAlleDokumenter(any()) } returns alleDocumenter
+            every { euxKlient.hentSed(any(), any()) } returns sedP8000recevied andThen sedP5000sent
+            every { euxKlient.hentSedDokumenter(any(), any()) } returns getResource("/pdf/pdfResponseUtenVedlegg.json")
+            every { personService.hentPerson(NorskIdent(fnr)) } returns createBrukerWith(fnr, "Lever", "Helt i live", "NOR", aktorId = aktoer)
+
+            val saker = listOf(
+                SakInformasjon(sakId = "34234234", sakType = YtelseType.ALDER, sakStatus = SakStatus.LOPENDE),
+                SakInformasjon(sakId = "23232312", sakType = YtelseType.UFOREP, sakStatus = SakStatus.AVSLUTTET),
+                SakInformasjon(sakId = "34234234234234", sakType = YtelseType.GENRL, sakStatus = SakStatus.LOPENDE)
+            )
+            every { fagmodulKlient.hentPensjonSaklist(aktoer) } returns saker
+
+            val meldingSlot = slot<String>()
+            every { oppgaveHandlerKafka.sendDefault(any(), capture(meldingSlot)).get() } returns mockk()
+            val (journalpost, journalpostResponse) = initJournalPostRequestSlot()
+            val hendelse = createHendelseJson(SedType.P6000, BucType.P_BUC_05)
+
+            listener.consumeSedSendt(hendelse, mockk(relaxed = true), mockk(relaxed = true))
+
+            val request = journalpost.captured
+            val oppgaveMelding = mapJsonToAny(meldingSlot.captured, typeRefs<OppgaveMelding>())
+
+            assertEquals("JOURNALFORING", oppgaveMelding.oppgaveType())
+            assertEquals(ID_OG_FORDELING, oppgaveMelding.tildeltEnhetsnr)
+            assertEquals(journalpostResponse.journalpostId, oppgaveMelding.journalpostId)
+
+            assertEquals("UTGAAENDE", request.journalpostType.name)
+            assertEquals(PENSJON, request.tema)
+            assertEquals(ID_OG_FORDELING, request.journalfoerendeEnhet)
+
+            verify(exactly = 1) { fagmodulKlient.hentAlleDokumenter(any()) }
+            verify(exactly = 1) { fagmodulKlient.hentPensjonSaklist(any()) }
+            verify(exactly = 2) { euxKlient.hentSed(any(), any()) }
+
+        }
+
+        @Test
+        fun `Scenario 13 - 4 Sed sendes som svar med fnr utland og sak finnes og er GENRL pa tidligere mottatt P8000, opprettes en journalføringsoppgave på tema PENSJON UTLAND`() {
+            val fnr = FNR_VOKSEN
+            val sakid = "1231232323"
+            val aktoer = "${fnr}111"
+            val sedP8000recevied = createSed(SedType.P8000, null, fdato = "1955-07-11")
+            val sedP5000sent = createSed(SedType.P5000, fnr, eessiSaknr = sakid)
+
+            val alleDocumenter = listOf(
+                Document("10001", SedType.P8000, DocStatus.RECEIVED),
+                Document("30002", SedType.P5000, DocStatus.SENT)
+            )
+
+            every { fagmodulKlient.hentAlleDokumenter(any()) } returns alleDocumenter
+            every { euxKlient.hentSed(any(), any()) } returns sedP8000recevied andThen sedP5000sent
+            every { euxKlient.hentSedDokumenter(any(), any()) } returns getResource("/pdf/pdfResponseUtenVedlegg.json")
+            every { personService.hentPerson(NorskIdent(fnr)) } returns createBrukerWith(fnr, "Lever", "Helt i live", "SWE", aktorId = aktoer)
+
+            val saker = listOf(
+                SakInformasjon(sakId = "34234234", sakType = YtelseType.ALDER, sakStatus = SakStatus.LOPENDE),
+                SakInformasjon(sakId = "23232312", sakType = YtelseType.UFOREP, sakStatus = SakStatus.AVSLUTTET),
+                SakInformasjon(sakId = "123123123123123123", sakType = YtelseType.GENRL, sakStatus = SakStatus.LOPENDE)
+            )
+            every { fagmodulKlient.hentPensjonSaklist(aktoer) } returns saker
+
+            val meldingSlot = slot<String>()
+            every { oppgaveHandlerKafka.sendDefault(any(), capture(meldingSlot)).get() } returns mockk()
+            val (journalpost, journalpostResponse) = initJournalPostRequestSlot()
+            val hendelse = createHendelseJson(SedType.P6000, BucType.P_BUC_05)
+
+            listener.consumeSedSendt(hendelse, mockk(relaxed = true), mockk(relaxed = true))
+
+            val request = journalpost.captured
+            val oppgaveMelding = mapJsonToAny(meldingSlot.captured, typeRefs<OppgaveMelding>())
+
+            assertEquals("JOURNALFORING", oppgaveMelding.oppgaveType())
+            assertEquals(ID_OG_FORDELING, oppgaveMelding.tildeltEnhetsnr)
+            assertEquals(journalpostResponse.journalpostId, oppgaveMelding.journalpostId)
+
+            assertEquals("UTGAAENDE", request.journalpostType.name)
+            assertEquals(PENSJON, request.tema)
+            assertEquals(ID_OG_FORDELING, request.journalfoerendeEnhet)
+
+            verify(exactly = 1) { fagmodulKlient.hentAlleDokumenter(any()) }
+            verify(exactly = 1) { fagmodulKlient.hentPensjonSaklist(any()) }
+            verify(exactly = 2) { euxKlient.hentSed(any(), any()) }
+
+        }
+
+        @Test
+        fun `Scenario 13 - 5 Sed sendes som svar med fnr og sak finnes og er UFOREP pa tidligere mottatt P8000, journalføres automatisk`() {
+            val fnr = FNR_VOKSEN
+            val sakid = "1231232323"
+            val aktoer = "${fnr}111"
+            val sedP8000recevied = createSed(SedType.P8000, null, fdato = "1955-07-11")
+            val sedP5000sent = createSed(SedType.P5000, fnr, eessiSaknr = sakid)
+
+            val alleDocumenter = listOf(
+                Document("10001", SedType.P8000, DocStatus.RECEIVED),
+                Document("30002", SedType.P5000, DocStatus.SENT)
+            )
+
+            every { fagmodulKlient.hentAlleDokumenter(any()) } returns alleDocumenter
+            every { euxKlient.hentSed(any(), any()) } returns sedP8000recevied andThen sedP5000sent
+            every { euxKlient.hentSedDokumenter(any(), any()) } returns getResource("/pdf/pdfResponseUtenVedlegg.json")
+            every { personService.hentPerson(NorskIdent(fnr)) } returns createBrukerWith(fnr, "Lever", "Helt i live", "NOR", aktorId = aktoer)
+
+            val saker = listOf(
+                SakInformasjon(sakId = "34234234", sakType = YtelseType.OMSORG, sakStatus = SakStatus.LOPENDE),
+                SakInformasjon(sakId = "23232312", sakType = YtelseType.GENRL, sakStatus = SakStatus.AVSLUTTET),
+                SakInformasjon(sakId = sakid, sakType = YtelseType.UFOREP, sakStatus = SakStatus.LOPENDE)
+            )
+            every { fagmodulKlient.hentPensjonSaklist(aktoer) } returns saker
+            every { journalpostKlient.oppdaterDistribusjonsinfo(any()) } returns Unit
+
+            val (journalpost, _) = initJournalPostRequestSlot(true)
+            val hendelse = createHendelseJson(SedType.P6000, BucType.P_BUC_05)
+
+            listener.consumeSedSendt(hendelse, mockk(relaxed = true), mockk(relaxed = true))
+
+            val request = journalpost.captured
+
+            assertEquals("UTGAAENDE", request.journalpostType.name)
+            assertEquals(UFORETRYGD, request.tema)
+            assertEquals(AUTOMATISK_JOURNALFORING, request.journalfoerendeEnhet)
+
+            verify(exactly = 1) { fagmodulKlient.hentAlleDokumenter(any()) }
+            verify(exactly = 1) { fagmodulKlient.hentPensjonSaklist(any()) }
+            verify(exactly = 2) { euxKlient.hentSed(any(), any()) }
+        }
+    }
+
+    /* ============================
+     * INNGÅENDE
+     * ============================ */
+
+    @Nested
+    @DisplayName("Inngående - Scenario 1")
+    inner class Scenario1Inngaende {
+        @Test
+        fun `Kun én person, mangler FNR`() {
+            testRunner(fnr = null, hendelseType = HendelseType.MOTTATT) {
+                assertEquals(PENSJON, it.tema)
+                assertEquals(ID_OG_FORDELING, it.journalfoerendeEnhet)
+            }
+        }
+
+        @Test
+        fun `Kun én person, ugyldig FNR`() {
+            testRunner(fnr = "1244091349018340918341029", hendelseType = HendelseType.MOTTATT) {
+                assertEquals(PENSJON, it.tema)
+                assertEquals(ID_OG_FORDELING, it.journalfoerendeEnhet)
+            }
+        }
+    }
+
+
+    @Nested
+    @DisplayName("Inngående - Scenario 2")
+    inner class Scenario2Inngaende {
+        @Test
+        fun `Manglende eller feil i FNR, DNR for forsikret - to personer angitt, ROLLE 03`() {
+            val sed = createSed(SedType.P8000, null, createAnnenPerson(fnr = FNR_BARN, rolle = Rolle.BARN), SAK_ID)
+            initCommonMocks(sed)
+
+            val barn = createBrukerWith(FNR_BARN, "Barn", "Vanlig", "NOR", "1213")
+            every { personService.hentPerson(NorskIdent(FNR_BARN)) } returns barn
+
+            val hendelse = createHendelseJson(SedType.P8000)
+
+            val meldingSlot = slot<String>()
+            every { oppgaveHandlerKafka.sendDefault(any(), capture(meldingSlot)).get() } returns mockk()
+
+            val (journalpost, _) = initJournalPostRequestSlot()
+
+            listener.consumeSedMottatt(hendelse, mockk(relaxed = true), mockk(relaxed = true))
+
+            val oppgaveMelding = mapJsonToAny(meldingSlot.captured, typeRefs<OppgaveMelding>())
+
+            assertEquals("JOURNALFORING", oppgaveMelding.oppgaveType())
+
+            val request = journalpost.captured
+            // forvent tema == PEN og enhet 9999
+            assertEquals(PENSJON, request.tema)
+            assertEquals(ID_OG_FORDELING, request.journalfoerendeEnhet)
+
+            verify(exactly = 1) { fagmodulKlient.hentAlleDokumenter(any()) }
+            verify(exactly = 1) { euxKlient.hentSed(any(), any()) }
+        }
+
+        @Test
+        fun `Manglende eller feil i FNR, DNR for forsikret - to personer angitt, ROLLE 02`() {
+            val sed = createSed(SedType.P8000, null, createAnnenPerson(fnr = FNR_VOKSEN, rolle = Rolle.FORSORGER), SAK_ID)
+            initCommonMocks(sed)
+
+            val voksen = createBrukerWith(FNR_VOKSEN, "Voksen", "Vanlig", "NOR", "1213")
+            every { personService.hentPerson(NorskIdent(FNR_VOKSEN)) } returns voksen
+
+            val hendelse = createHendelseJson(SedType.P8000)
+
+            val meldingSlot = slot<String>()
+            every { oppgaveHandlerKafka.sendDefault(any(), capture(meldingSlot)).get() } returns mockk()
+
+            val (journalpost, _) = initJournalPostRequestSlot()
+
+            listener.consumeSedMottatt(hendelse, mockk(relaxed = true), mockk(relaxed = true))
+
+            val oppgaveMelding = mapJsonToAny(meldingSlot.captured, typeRefs<OppgaveMelding>())
+
+            assertEquals("JOURNALFORING", oppgaveMelding.oppgaveType())
+
+            val request = journalpost.captured
+            assertEquals(PENSJON, request.tema)
+            assertEquals(ID_OG_FORDELING, request.journalfoerendeEnhet)
+
+            verify(exactly = 1) { fagmodulKlient.hentAlleDokumenter(any()) }
+            verify(exactly = 1) { euxKlient.hentSed(any(), any()) }
+        }
+    }
+
+    @Nested
+    @DisplayName("Inngående - Scenario 3")
+    inner class Scenario3Inngaende {
+        @Test
+        fun `Manglende eller feil FNR-DNR - to personer angitt - etterlatte`() {
+            val sed = createSed(SedType.P8000, null, createAnnenPerson(fnr = FNR_VOKSEN, rolle = Rolle.ETTERLATTE), SAK_ID)
+            initCommonMocks(sed)
+
+            val voksen = createBrukerWith(FNR_VOKSEN, "Voksen", "Vanlig", "NOR", "1213")
+            every { personService.hentPerson(NorskIdent(FNR_VOKSEN)) } returns voksen
+
+            val hendelse = createHendelseJson(SedType.P8000)
+
+            val meldingSlot = slot<String>()
+            every { oppgaveHandlerKafka.sendDefault(any(), capture(meldingSlot)).get() } returns mockk()
+
+            val (journalpost, _) = initJournalPostRequestSlot()
+
+            listener.consumeSedMottatt(hendelse, mockk(relaxed = true), mockk(relaxed = true))
+
+            val oppgaveMelding = mapJsonToAny(meldingSlot.captured, typeRefs<OppgaveMelding>())
+
+            assertEquals("JOURNALFORING", oppgaveMelding.oppgaveType())
+
+            val request = journalpost.captured
+            // forvent tema == PEN og enhet 9999
+            assertEquals(PENSJON, request.tema)
+            assertEquals(NFP_UTLAND_AALESUND, request.journalfoerendeEnhet)
+
+            verify(exactly = 1) { fagmodulKlient.hentAlleDokumenter(any()) }
+            verify(exactly = 1) { euxKlient.hentSed(any(), any()) }
+        }
+
+        @Test
+        fun `Manglende eller feil FNR-DNR - to personer angitt - etterlatte med feil FNR for annen person, eller soker`() {
+            val sed = createSed(SedType.P8000, FNR_VOKSEN_2, createAnnenPerson(fnr = null, rolle = Rolle.ETTERLATTE), SAK_ID)
+            initCommonMocks(sed)
+
+            val voksen = createBrukerWith(FNR_VOKSEN_2, "Voksen", "Vanlig", "NOR", "1213")
+            every { personService.hentPerson(NorskIdent(FNR_VOKSEN_2)) } returns voksen
+
+            val hendelse = createHendelseJson(SedType.P8000)
+            val meldingSlot = slot<String>()
+            every { oppgaveHandlerKafka.sendDefault(any(), capture(meldingSlot)).get() } returns mockk()
+
+            val (journalpost, _) = initJournalPostRequestSlot()
+
+            listener.consumeSedMottatt(hendelse, mockk(relaxed = true), mockk(relaxed = true))
+
+            val oppgaveMelding = mapJsonToAny(meldingSlot.captured, typeRefs<OppgaveMelding>())
+
+            assertEquals("JOURNALFORING", oppgaveMelding.oppgaveType())
+
+            val request = journalpost.captured
+            // forvent tema == PEN og enhet 9999
+            assertEquals(PENSJON, request.tema)
+            assertEquals(ID_OG_FORDELING, request.journalfoerendeEnhet)
+
+            verify(exactly = 1) { fagmodulKlient.hentAlleDokumenter(any()) }
+            verify(exactly = 1) { euxKlient.hentSed(any(), any()) }
+        }
+    }
+
+
+    @Nested
+    @DisplayName("Inngående - Scenario 4")
+    inner class Scenario4Inngaende{
+        @Test
+        fun `Én person, gyldig fnr`() {
+            testRunner(FNR_OVER_60, sakId = null, hendelseType = HendelseType.MOTTATT) {
+                assertEquals(PENSJON, it.tema)
+                assertEquals(NFP_UTLAND_AALESUND, it.journalfoerendeEnhet)
+            }
+
+            testRunner(FNR_VOKSEN, sakId = null, hendelseType = HendelseType.MOTTATT) {
+                assertEquals(PENSJON, it.tema)
+                assertEquals(UFORE_UTLANDSTILSNITT, it.journalfoerendeEnhet)
+            }
+
+            testRunner(FNR_BARN, sakId = null, hendelseType = HendelseType.MOTTATT) {
+                assertEquals(PENSJON, it.tema)
+                assertEquals(NFP_UTLAND_AALESUND, it.journalfoerendeEnhet)
+            }
+        }
+
+        @Test
+        fun `Én person, gyldig fnr, bosatt utland`() {
+            testRunner(FNR_OVER_60, sakId = null, hendelseType = HendelseType.MOTTATT, land = "SWE") {
+                assertEquals(PENSJON, it.tema)
+                assertEquals(PENSJON_UTLAND, it.journalfoerendeEnhet)
+            }
+
+            testRunner(FNR_VOKSEN, sakId = null, hendelseType = HendelseType.MOTTATT, land = "SWE") {
+                assertEquals(PENSJON, it.tema)
+                assertEquals(UFORE_UTLAND, it.journalfoerendeEnhet)
+            }
+
+            testRunner(FNR_BARN, sakId = null, hendelseType = HendelseType.MOTTATT, land = "SWE") {
+                assertEquals(PENSJON, it.tema)
+                assertEquals(PENSJON_UTLAND, it.journalfoerendeEnhet)
+            }
+        }
+    }
+
+
+    @Nested
+    @DisplayName("Inngående - Scenario 5")
+    inner class Scenario5Inngaende {
+        @Test
+        fun `To personer angitt, gyldig fnr, rolle er 03, bosatt norge`() {
+            testRunnerFlerePersoner(FNR_OVER_60, FNR_VOKSEN, rolle = Rolle.BARN, sakId = null, hendelseType = HendelseType.MOTTATT) {
+                assertEquals(PENSJON, it.tema)
+                assertEquals(NFP_UTLAND_AALESUND, it.journalfoerendeEnhet)
+                assertEquals(FNR_OVER_60, it.bruker!!.id)
+            }
+
+            testRunnerFlerePersoner(FNR_VOKSEN, FNR_BARN, rolle = Rolle.BARN, sakId = null, hendelseType = HendelseType.MOTTATT) {
+                assertEquals(PENSJON, it.tema)
+                assertEquals(UFORE_UTLANDSTILSNITT, it.journalfoerendeEnhet)
+                assertEquals(FNR_VOKSEN, it.bruker!!.id)
+            }
+        }
+
+        @Test
+        fun `To personer angitt, gyldig fnr, rolle er 02, bosatt norge`() {
+            testRunnerFlerePersoner(FNR_OVER_60, FNR_VOKSEN, rolle = Rolle.FORSORGER, sakId = null, hendelseType = HendelseType.MOTTATT) {
+                assertEquals(PENSJON, it.tema)
+                assertEquals(NFP_UTLAND_AALESUND, it.journalfoerendeEnhet)
+                assertEquals(FNR_OVER_60, it.bruker!!.id)
+            }
+
+            testRunnerFlerePersoner(FNR_VOKSEN, FNR_BARN, rolle = Rolle.FORSORGER, sakId = null, hendelseType = HendelseType.MOTTATT) {
+                assertEquals(PENSJON, it.tema)
+                assertEquals(UFORE_UTLANDSTILSNITT, it.journalfoerendeEnhet)
+                assertEquals(FNR_VOKSEN, it.bruker!!.id)
+            }
+        }
+
+        @Test
+        fun `To personer angitt, gyldig fnr, rolle er 03, bosatt utland`() {
+            testRunnerFlerePersoner(FNR_OVER_60, FNR_VOKSEN, rolle = Rolle.BARN, sakId = null, hendelseType = HendelseType.MOTTATT, land = "SWE") {
+                assertEquals(PENSJON, it.tema)
+                assertEquals(PENSJON_UTLAND, it.journalfoerendeEnhet)
+                assertEquals(FNR_OVER_60, it.bruker!!.id)
+            }
+
+            testRunnerFlerePersoner(FNR_VOKSEN, FNR_BARN, rolle = Rolle.BARN, sakId = null, hendelseType = HendelseType.MOTTATT, land = "SWE") {
+                assertEquals(PENSJON, it.tema)
+                assertEquals(UFORE_UTLAND, it.journalfoerendeEnhet)
+                assertEquals(FNR_VOKSEN, it.bruker!!.id)
+            }
+        }
+
+        @Test
+        fun `To personer angitt, gyldig fnr, rolle er 02, bosatt utland`() {
+            testRunnerFlerePersoner(FNR_OVER_60, FNR_VOKSEN, rolle = Rolle.FORSORGER, sakId = null, hendelseType = HendelseType.MOTTATT, land = "SWE") {
+                assertEquals(PENSJON, it.tema)
+                assertEquals(PENSJON_UTLAND, it.journalfoerendeEnhet)
+                assertEquals(FNR_OVER_60, it.bruker!!.id)
+            }
+
+            testRunnerFlerePersoner(FNR_VOKSEN, FNR_BARN, rolle = Rolle.FORSORGER, sakId = null, hendelseType = HendelseType.MOTTATT, land = "SWE") {
+                assertEquals(PENSJON, it.tema)
+                assertEquals(UFORE_UTLAND, it.journalfoerendeEnhet)
+                assertEquals(FNR_VOKSEN, it.bruker!!.id)
+            }
+        }
+
+        @Test
+        fun `To personer angitt, gyldig fnr og ugyldig annenperson, rolle er 02, bosatt utland del 1`() {
+            val sed = createSed(SedType.P8000, FNR_OVER_60, createAnnenPerson(fnr = FNR_BARN, rolle = Rolle.FORSORGER), null)
+            initCommonMocks(sed)
+
+            val voksen = createBrukerWith(FNR_OVER_60, "Voksen", "Vanlig", "SWE", "1213")
+            every { personService.hentPerson(NorskIdent(FNR_OVER_60)) } returns voksen
+            every { personService.hentPerson(NorskIdent(FNR_BARN)) } returns null
+
+            val hendelse = createHendelseJson(SedType.P8000)
+
+            val meldingSlot = slot<String>()
+            every { oppgaveHandlerKafka.sendDefault(any(), capture(meldingSlot)).get() } returns mockk()
+
+            val (journalpost, _) = initJournalPostRequestSlot()
+
+            listener.consumeSedMottatt(hendelse, mockk(relaxed = true), mockk(relaxed = true))
+
+            val oppgaveMelding = mapJsonToAny(meldingSlot.captured, typeRefs<OppgaveMelding>())
+
+            assertEquals("JOURNALFORING", oppgaveMelding.oppgaveType())
+
+            val request = journalpost.captured
+            assertEquals(PENSJON, request.tema)
+            assertEquals(PENSJON_UTLAND, request.journalfoerendeEnhet)
+            assertEquals(FNR_OVER_60, request.bruker!!.id)
+
+            verify(exactly = 1) { fagmodulKlient.hentAlleDokumenter(any()) }
+            verify(exactly = 1) { euxKlient.hentSed(any(), any()) }
+        }
+
+        @Test
+        fun `To personer angitt, gyldig fnr og ufgyldig fnr annenperson, rolle er 02, bosatt utland del 2`() {
+            val valgtFNR = FNR_VOKSEN
+            val sed = createSed(SedType.P8000, valgtFNR, createAnnenPerson(fnr = FNR_BARN, rolle = Rolle.FORSORGER), null)
+            initCommonMocks(sed)
+
+            val voksen = createBrukerWith(valgtFNR, "Voksen", "Vanlig", "SWE", "1213")
+            every { personService.hentPerson(NorskIdent(valgtFNR)) } returns voksen
+            every { personService.hentPerson(NorskIdent(FNR_BARN)) } returns null
+
+            val hendelse = createHendelseJson(SedType.P8000)
+
+            val meldingSlot = slot<String>()
+            every { oppgaveHandlerKafka.sendDefault(any(), capture(meldingSlot)).get() } returns mockk()
+
+            val (journalpost, _) = initJournalPostRequestSlot()
+
+            listener.consumeSedMottatt(hendelse, mockk(relaxed = true), mockk(relaxed = true))
+
+            val oppgaveMelding = mapJsonToAny(meldingSlot.captured, typeRefs<OppgaveMelding>())
+
+            assertEquals("JOURNALFORING", oppgaveMelding.oppgaveType())
+
+            val request = journalpost.captured
+            assertEquals(PENSJON, request.tema)
+            assertEquals(UFORE_UTLAND, request.journalfoerendeEnhet)
+            assertEquals(valgtFNR, request.bruker!!.id)
+
+            verify(exactly = 1) { fagmodulKlient.hentAlleDokumenter(any()) }
+            verify(exactly = 1) { euxKlient.hentSed(any(), any()) }
+        }
+
+        @Test
+        fun `To personer angitt, gyldig fnr og ufgyldig fnr annenperson, rolle er 02, bosatt Norge del 3`() {
+            val valgtFNR = FNR_VOKSEN
+            val sed = createSed(SedType.P8000, valgtFNR, createAnnenPerson(fnr = FNR_BARN, rolle = Rolle.FORSORGER), null)
+            initCommonMocks(sed)
+
+            val voksen = createBrukerWith(valgtFNR, "Voksen", "Vanlig", "NOR", "1213")
+            every { personService.hentPerson(NorskIdent(valgtFNR)) } returns voksen
+            every { personService.hentPerson(NorskIdent(FNR_BARN)) } returns null
+
+            val hendelse = createHendelseJson(SedType.P8000)
+
+            val meldingSlot = slot<String>()
+            every { oppgaveHandlerKafka.sendDefault(any(), capture(meldingSlot)).get() } returns mockk()
+
+            val (journalpost, _) = initJournalPostRequestSlot()
+
+            listener.consumeSedMottatt(hendelse, mockk(relaxed = true), mockk(relaxed = true))
+
+            val oppgaveMelding = mapJsonToAny(meldingSlot.captured, typeRefs<OppgaveMelding>())
+
+            assertEquals("JOURNALFORING", oppgaveMelding.oppgaveType())
+
+            val request = journalpost.captured
+            assertEquals(PENSJON, request.tema)
+            assertEquals(UFORE_UTLANDSTILSNITT, request.journalfoerendeEnhet)
+            assertEquals(valgtFNR, request.bruker!!.id)
+
+            verify(exactly = 1) { fagmodulKlient.hentAlleDokumenter(any()) }
+            verify(exactly = 1) { euxKlient.hentSed(any(), any()) }
+        }
+
+        @Test
+        fun `To personer angitt, gyldig fnr og ufgyldig fnr annenperson, rolle er 02, bosatt Norge del 4`() {
+            val valgtFNR = FNR_OVER_60
+            val sed = createSed(SedType.P8000, valgtFNR, createAnnenPerson(fnr = FNR_BARN, rolle = Rolle.FORSORGER), null)
+            initCommonMocks(sed)
+
+            val voksen = createBrukerWith(valgtFNR, "Voksen", "Vanlig", "NOR", "1213")
+            every { personService.hentPerson(NorskIdent(valgtFNR)) } returns voksen
+            every { personService.hentPerson(NorskIdent(FNR_BARN)) } returns null
+
+            val hendelse = createHendelseJson(SedType.P8000)
+
+            val meldingSlot = slot<String>()
+            every { oppgaveHandlerKafka.sendDefault(any(), capture(meldingSlot)).get() } returns mockk()
+
+            val (journalpost, _) = initJournalPostRequestSlot()
+
+            listener.consumeSedMottatt(hendelse, mockk(relaxed = true), mockk(relaxed = true))
+
+            val oppgaveMelding = mapJsonToAny(meldingSlot.captured, typeRefs<OppgaveMelding>())
+
+            assertEquals("JOURNALFORING", oppgaveMelding.oppgaveType())
+
+            val request = journalpost.captured
+            assertEquals(PENSJON, request.tema)
+            assertEquals(NFP_UTLAND_AALESUND, request.journalfoerendeEnhet)
+            assertEquals(valgtFNR, request.bruker!!.id)
+
+            verify(exactly = 1) { fagmodulKlient.hentAlleDokumenter(any()) }
+            verify(exactly = 1) { euxKlient.hentSed(any(), any()) }
+        }
+
+        @Test
+        fun `To personer angitt, gyldig fnr og ufgyldig fnr annenperson, rolle er 01, bosatt Norge del 4`() {
+            val valgtFNR = FNR_OVER_60
+            val sed = createSed(SedType.P8000, valgtFNR, createAnnenPerson(fnr = FNR_BARN, rolle = Rolle.ETTERLATTE), null)
+            initCommonMocks(sed)
+
+            val voksen = createBrukerWith(valgtFNR, "Voksen", "Vanlig", "NOR", "1213")
+            every { personService.hentPerson(NorskIdent(valgtFNR)) } returns voksen
+            every { personService.hentPerson(NorskIdent(FNR_BARN)) } returns null
+
+            val hendelse = createHendelseJson(SedType.P8000)
+
+            val meldingSlot = slot<String>()
+            every { oppgaveHandlerKafka.sendDefault(any(), capture(meldingSlot)).get() } returns mockk()
+
+            val (journalpost, _) = initJournalPostRequestSlot()
+
+            listener.consumeSedMottatt(hendelse, mockk(relaxed = true), mockk(relaxed = true))
+
+            val oppgaveMelding = mapJsonToAny(meldingSlot.captured, typeRefs<OppgaveMelding>())
+
+            assertEquals("JOURNALFORING", oppgaveMelding.oppgaveType())
+
+            val request = journalpost.captured
+            assertEquals(PENSJON, request.tema)
+            assertEquals(ID_OG_FORDELING, request.journalfoerendeEnhet)
+            Assertions.assertNull(request.bruker)
+
+            verify(exactly = 1) { fagmodulKlient.hentAlleDokumenter(any()) }
+            verify(exactly = 1) { euxKlient.hentSed(any(), any()) }
+        }
+
+    }
+
+
+    @Nested
+    @DisplayName("Inngående - Scenario 6")
+    inner class Scenario6Inngaende {
+        @Test
+        fun `To personer angitt, gyldig fnr, rolle 02 etterlatte, bosatt norge`() {
+            testRunnerFlerePersoner(FNR_OVER_60, FNR_BARN, rolle = Rolle.ETTERLATTE, sakId = null, hendelseType = HendelseType.MOTTATT) {
+                assertEquals(PENSJON, it.tema)
+                assertEquals(NFP_UTLAND_AALESUND, it.journalfoerendeEnhet)
+            }
+
+            testRunnerFlerePersoner(FNR_OVER_60, FNR_VOKSEN, rolle = Rolle.ETTERLATTE, sakId = null, hendelseType = HendelseType.MOTTATT) {
+                assertEquals(PENSJON, it.tema)
+                assertEquals(NFP_UTLAND_AALESUND, it.journalfoerendeEnhet)
+            }
+        }
+
+        @Test
+        fun `To personer angitt, gyldig fnr, rolle 02 etterlatte, bosatt utland`() {
+            testRunnerFlerePersoner(FNR_OVER_60, FNR_BARN, rolle = Rolle.ETTERLATTE, sakId = null, hendelseType = HendelseType.MOTTATT, land = "SWE") {
+                assertEquals(PENSJON, it.tema)
+                assertEquals(PENSJON_UTLAND, it.journalfoerendeEnhet)
+            }
+
+            testRunnerFlerePersoner(FNR_OVER_60, FNR_VOKSEN, rolle = Rolle.ETTERLATTE, sakId = null, hendelseType = HendelseType.MOTTATT, land = "SWE") {
+                assertEquals(PENSJON, it.tema)
+                assertEquals(PENSJON_UTLAND, it.journalfoerendeEnhet)
+            }
+        }
+    }
+
+    private fun initCommonMocks(sed: SED) {
+        val documents = mapJsonToAny(getResource("/fagmodul/alldocumentsids.json"), typeRefs<List<Document>>())
+
+        every { fagmodulKlient.hentAlleDokumenter(any()) } returns documents
+        every { euxKlient.hentSed(any(), any()) } returns sed
+        every { euxKlient.hentSedDokumenter(any(), any()) } returns getResource("/pdf/pdfResponseUtenVedlegg.json")
+    }
+
+    private fun getResource(resourcePath: String): String =
+        javaClass.getResource(resourcePath).readText()
 }
